Add address type enum, document address layout, change function names
The different address types are now explicitly enumerated
so that we don't need to use magic numbers elsewhere to
refer to the different address types.

The way in which the address array is used for each address
type is now documented in hash_address.h.

New function names make it easier to see which functions should
be used for what kind of address.