Hash addresses: Split layer and node index for SPHINCS addresses; add getters
The layer and node index of a SPHINCS address can now be accessed
separately, which makes it much easier to address nodes in the
SPHINCS tree.
This also adds a bunch of getters, so that code outside of this module
does not need to know about the inner structure of the addressing scheme.