[Tests failing] Change address handling in sign_leaf and verify_leaf
Previously the function expected a start and an end layer which would
define from where to where the signature was created or verified.
Now the functions respect the layer that is set in the passed address,
and only accept a parameter that tells them how many layers should be
signed/verified starting from that layer.