Fix how the address of a tree's parent is calculated
This bug subtly lead to wrong addresses throughout the signing process.
Since I messed up the calculation of the parent address _consistently_
throughout the code, the bug didn't really show up until I tried to
use the batch API.

Because of this bug, the upper part of the signature (say from level 1 to level
N_LEVELS - 1) would change, depending on which leaf was chosen on level 0. This
is fine if you use the same leaf for signing and verifying, but the batch API
doesn't do that; in the batch API the upper part of the signature is
initialized for a random subtree on level 0, and later on signatures are
created for different leaves of that subtree on level 0.