Allow user to pass explicit subtree index or generate random index
When calling crypto_context_init, a user can now provide the index of the
subtree they would like to use, or pass -1 to have a random index
generated. The random index is currently based on the seed in the secret key
and SEED_BYTES read from dev/urandom.

In the future the randomly generated index should use getentropy instead if
available.