Move code, header, and tests for batch API to separate module
This separates the code for classic Sphincs from the code for sequential batch
signatures (the one and only variant of batch signatures that is currently
implemented.

There is now also a separate header file for the batch api, and the tests
for the batch api also happen in a separate file.