Change sequential batch signatures
Previously, the STS would be initialized with a random subtree index, and the
leaf nodes of that subtree would be used from left to right to sign messages.
However, with that approach, there's a risk that left leaf nodes in the subtree
see more usage than other leaf nodes. Whenever a STS is not utilized 100%, leaf
nodes to the right of the subtree would remain unused. That would lead to an
imbalanced usage, possibly more than 8 re-uses of some HORST key pairs, and
thereby possibly breaking the scheme.

This commit changes the approach to fix that. Now, the STS is initialized with
a particular leaf index, pointing to a HORST key pair. In addition to that, the
STS keeps track of the number of messages that have been signed with that STS.
All signatures will still be made with HORST key pairs of the same subtree, but
instead of starting with the first leaf node and continuing to the last, we
start at a random leaf node, and continue to the right, potentially wrapping
back to the first leaf node of the subtree.

This new approach leads to a more even distribution of signatures over the leaf
nodes.