Add struct for subtree signature and inline crypto_sts_update
We don't use crypto_sts_update any more, it's not clear what use case
it would have in the public API, and it's complicating things unnecessarily.

This also adds a struct for the signature, similar to what we already have
for sequential batch signing. The struct isn't fully in use yet, we merely
compare the addresses of the struct to what's happening in the code, and will
switch to the struct in the next couple of commmits.